+++ src/genai/streamlit_pages/__init__.py
@@ -1,0 +1,2 @@
+from .eli3_page import eli3
+from .eyfs_page import early_year_activity_plan
