+++ src/genai/streamlit_pages/eli3_page.py
@@ -1,0 +1,41 @@
+import streamlit as st
+
+from genai.eli3 import TextGenerator
+
+
+def eli3() -> None:
+    """Explain me a concept like I'm 3."""
+    st.title("Explain like I am a three year old")
+
+    # Create the generator
+    selected_model = st.radio(label="**OpenAI model**", options=["gpt-3.5-turbo", "gpt-4"])
+    try:
+        generator = TextGenerator(
+            path="src/genai/eli3/prompts/eli3.json",
+            model_name=selected_model,
+        )
+    except Exception:  # Dirty hack to work with local secrets and not break the app on Streamlit Share
+        generator = TextGenerator(
+            api_key=st.secret("OPENAI_API_KEY"),
+            path="src/genai/eli3/prompts/eli3.json",
+            model_name=selected_model,
+        )
+
+    prompt_selector = st.radio(label="**Generate with custom prompt**", options=["Default", "Custom"])
+
+    if prompt_selector == "Custom":
+        prompt = st.text_area("Write your own prompt", value=generator.prompt_template.template)
+    else:
+        prompt = None
+
+    # Get the user input
+    question = st.text_input(
+        label="**Question**",
+        value="How can whales breath in water?",
+        help="Ask the large language model a question.",
+    )
+
+    # Generate the answer
+    if st.button(label="**Generate**", help="Generate an answer."):
+        answer = generator.generate(question, prompt=prompt)
+        st.write(answer)
